+++ db/migrate/20160803194259_add_shadow_banned_to_users.rb
@@ -1,0 +1,6 @@
+class AddShadowBannedToUsers < ActiveRecord::Migration[5.0]
+  def change
+    add_column :users, :shadow_banned, :boolean
+    add_index :users, :shadow_banned
+  end
+end