+++ test/controllers/chapters_controller_test.rb
@@ -6,7 +6,7 @@ require 'test_helper'
 class ChaptersControllerTest < ActionController::TestCase
   setup do
     @chapter = chapters(:one)
-    login(users(:user_1))
+    @regular_user = users(:user_1)
   end
 
   def chapter_params
@@ -19,17 +19,20 @@ class ChaptersControllerTest < ActionController::TestCase
   end
 
   test 'should get index' do
+    login(@regular_user)
     get :index
     assert_response :success
     assert_not_nil assigns(:chapters)
   end
 
   test 'should get new' do
+    login(@regular_user)
     get :new
     assert_response :success
   end
 
   test 'should create chapter' do
+    login(@regular_user)
     assert_difference('Chapter.count') do
       post :create, params: { chapter: chapter_params }
     end
@@ -41,6 +44,7 @@ class ChaptersControllerTest < ActionController::TestCase
   end
 
   test 'should not create chapter with blank title' do
+    login(@regular_user)
     assert_difference('Chapter.count', 0) do
       post :create, params: { chapter: chapter_params.merge(title: '') }
     end
@@ -48,28 +52,59 @@ class ChaptersControllerTest < ActionController::TestCase
     assert_response :success
   end
 
+  test 'should shadow ban new spammer' do
+    login(users(:new_spammer))
+    assert_difference('Chapter.count') do
+      post :create, params: { chapter: { title: 'http://www.example.com', slug: 'http-www-example-com', description: 'http://www.example.com' } }
+    end
+    assert_equal true, assigns(:chapter).user.shadow_banned
+    assert_redirected_to assigns(:chapter)
+  end
+
+  test 'should not shadow ban verified user' do
+    login(users(:verified_user))
+    assert_difference('Chapter.count') do
+      post :create, params: { chapter: { title: 'http://www.example.com', slug: 'http-www-example-com', description: 'http://www.example.com' } }
+    end
+    assert_equal false, assigns(:chapter).user.shadow_banned
+    assert_redirected_to assigns(:chapter)
+  end
+
+  # test 'should not create chapter for shadow banned user' do
+  #   login(users(:shadow_banned))
+  #   assert_difference('Chapter.count', 0) do
+  #     post :create, params: { chapter: chapter_params.merge(title: '') }
+  #   end
+  #   assert_redirected_to chapters_path
+  # end
+
   test 'should show chapter' do
+    login(@regular_user)
     get :show, params: { id: @chapter }
     assert_response :success
   end
 
   test 'should get edit' do
+    login(@regular_user)
     get :edit, params: { id: @chapter }
     assert_response :success
   end
 
   test 'should update chapter' do
+    login(@regular_user)
     patch :update, params: { id: @chapter, chapter: chapter_params }
     assert_redirected_to assigns(:chapter)
   end
 
   test 'should not update chapter with blank title' do
+    login(@regular_user)
     patch :update, params: { id: @chapter, chapter: chapter_params.merge(title: '') }
     assert_template 'edit'
     assert_response :success
   end
 
   test 'should destroy chapter' do
+    login(@regular_user)
     assert_difference('Chapter.current.count', -1) do
       delete :destroy, params: { id: @chapter }
     end